update metadata section component to be more generic
